move caseDatatype
