+++ src/main/java/uk/gov/digital/ho/hocs/casework/api/dto/CreateCaseNoteRequest.java
@@ -3,12 +3,8 @@ package uk.gov.digital.ho.hocs.casework.api.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import uk.gov.digital.ho.hocs.casework.domain.model.CaseDataType;
 
 import javax.validation.constraints.NotEmpty;
-import java.time.LocalDate;
-import java.util.Map;
-import java.util.UUID;
 
 @AllArgsConstructor
 @Getter
