+++ src/test/java/uk/gov/digital/ho/hocs/casework/api/CaseDataTimelineCaseIntegrationTest.java
@@ -21,7 +21,7 @@ import uk.gov.digital.ho.hocs.casework.client.auditclient.dto.GetAuditListRespon
 import uk.gov.digital.ho.hocs.casework.client.auditclient.dto.GetAuditResponse;
 import uk.gov.digital.ho.hocs.casework.client.infoclient.PermissionDto;
 import uk.gov.digital.ho.hocs.casework.client.infoclient.TeamDto;
-import uk.gov.digital.ho.hocs.casework.domain.model.CaseDataType;
+import uk.gov.digital.ho.hocs.casework.api.dto.CaseDataType;
 import uk.gov.digital.ho.hocs.casework.security.AccessLevel;
 
 import java.io.IOException;
