+++ src/test/java/uk/gov/digital/ho/hocs/casework/api/dto/CreateCaseNoteRequestTest.java
@@ -1,11 +1,6 @@
 package uk.gov.digital.ho.hocs.casework.api.dto;
 
 import org.junit.Test;
-import uk.gov.digital.ho.hocs.casework.domain.model.CaseDataType;
-
-import java.time.LocalDate;
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
